+++ components/groups.js
@@ -1,3 +1,4 @@
+'use client';
 import { useEffect, useState } from 'react';
 import axios from 'axios';
 import _ from 'lodash';
