+++ next.config.js
@@ -1,0 +1,8 @@
+const nextConfig = {
+    experimental: {
+        appDir: true,
+        // https://beta.nextjs.org/docs/api-reference/next-config#servercomponentsexternalpackages
+        // serverComponentsExternalPackages: ['react-bootstrap'],
+    },
+};
+module.exports = nextConfig;
