+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "strawberry-scim",
-  "version": "1.2.0",
+  "version": "1.3.0",
   "private": false,
   "description": "Lightweight SCIM client with simple Bootstrap UI to make HTTP requests to your SCIM Endpoints. Based mainly on AAD SCIM client.",
   "displayName": "🍓 SCIM client",
