--- pages/_app.js
@@ -1,8 +1,0 @@
-import '../styles/global.css';
-import { ToastProvider } from '../components/toast-manager';
-
-export default function App({ Component, pageProps }) {
-    return <ToastProvider>
-        <Component {...pageProps} />
-    </ToastProvider>;
-}
