--- pages/_document.js
@@ -1,29 +1,0 @@
-import Document, { Html, Head, Main, NextScript } from 'next/document';
-import pjson from '../package.json';
-
-class MyDocument extends Document {
-    static async getInitialProps(ctx) {
-        const initialProps = await Document.getInitialProps(ctx);
-        return { ...initialProps };
-    }
-
-    render() {
-        return (
-            <Html>
-                <Head>
-                    <meta name="description" content={pjson.description} />
-                    <title>{[pjson.displayName].join(' - ')}</title>
-                    <link rel='icon' href='/favicon.ico' />
-                    <link href='https://cdn.jsdelivr.net/npm/bootstrap@5.0.0-beta3/dist/css/bootstrap.min.css' rel='stylesheet' integrity='sha384-eOJMYsd53ii+scO/bJGFsiCZc+5NDVN2yr8+0RDqr0Ql0h+rP48ckxlpbzKgwra6' crossOrigin='anonymous' />
-                    <script src='https://cdn.jsdelivr.net/npm/bootstrap@5.0.0-beta3/dist/js/bootstrap.bundle.min.js' integrity='sha384-JEW9xMcG8R+pH31jmWH6WWP0WintQrMb4s7ZOdauHnUtxwoG2vI5DkLtS3qm9Ekf' crossOrigin='anonymous' />
-                </Head>
-                <body>
-                    <Main />
-                    <NextScript />
-                </body>
-            </Html>
-        );
-    }
-}
-
-export default MyDocument;
